Fix: Retrieve the frontend code of the extractor operator for the Python version.
