+++ web/src/pages/agent/utils/tests/extractor-transform.test.ts
@@ -1,7 +1,17 @@
 import { transformExtractorConfigToForm } from '@/utils/pipeline-operator';
 import { transformExtractorParams } from '../../utils';
 
+let mockIsGoBackend = true;
+jest.mock('@/utils/backend-runtime', () => ({
+  isGoBackend: () => mockIsGoBackend,
+  getBackendLanguage: () => (mockIsGoBackend ? 'go' : 'python'),
+}));
+
 describe('Extractor parameter transformations & precedence', () => {
+  beforeEach(() => {
+    mockIsGoBackend = true;
+  });
+
   describe('transformExtractorParams', () => {
     it('synchronizes nested modular configs to flat fields and preserves nested objects', () => {
       const input: any = {
@@ -110,6 +120,54 @@ describe('Extractor parameter transformations & precedence', () => {
       expect(result.keywords.top_n).toBe(4);
       expect(result.questions.top_n).toBe(2);
       expect(result.tags.top_n).toBe(1);
+    });
+  });
+
+  describe('python backend keeps the legacy flat shape', () => {
+    beforeEach(() => {
+      mockIsGoBackend = false;
+    });
+
+    it('transformExtractorParams only wraps prompts, without nested configs', () => {
+      const input: any = {
+        field_name: 'summary',
+        sys_prompt: 'sys',
+        prompts: 'user prompt',
+        auto_keywords: 3,
+        auto_questions: 2,
+        auto_tags: 1,
+        tag_file_id: 'tag-1',
+        llm_id: 'gpt-4',
+      };
+
+      const result = transformExtractorParams(input);
+
+      expect(result).toEqual({
+        ...input,
+        prompts: [{ content: 'user prompt', role: 'user' }],
+      });
+      expect(result).not.toHaveProperty('keywords');
+      expect(result).not.toHaveProperty('questions');
+      expect(result).not.toHaveProperty('tags');
+      expect(result).not.toHaveProperty('summary');
+      expect(result).not.toHaveProperty('metadata_config');
+    });
+
+    it('transformExtractorConfigToForm only unwraps prompts, without nested configs', () => {
+      const config = {
+        prompts: [{ content: 'user prompt', role: 'user' }],
+        auto_keywords: 4,
+        auto_tags: 1,
+      };
+
+      const result = transformExtractorConfigToForm(config);
+
+      expect(result.prompts).toBe('user prompt');
+      expect(result).not.toHaveProperty('keywords');
+      expect(result).not.toHaveProperty('summary');
+      expect(result).not.toHaveProperty('metadata_config');
+      expect(result).not.toHaveProperty('enable_summary');
+      expect(result).not.toHaveProperty('enable_metadata');
     });
   });
 });
